Conflicts:
	CHANGELOG.md

+++ app/com/lynxanalytics/biggraph/controllers/CleanerController.scala
@@ -26,7 +26,7 @@ case class DataFilesStatus(
   total: DataFilesStats,
   methods: List[DataFilesStats])
 
-case class Method(
+case class CleanerMethod(
   id: String,
   name: String,
   desc: String,
@@ -35,6 +35,7 @@ case class Method(
 case class MarkDeletedRequest(method: String)
 
 case class AllFiles(
+  partitioned: Map[String, Long],
   entities: Map[String, Long],
   operations: Map[String, Long],
   scalars: Map[String, Long])
@@ -43,20 +44,20 @@ class CleanerController(environment: BigGraphEnvironment) {
   implicit val manager = environment.metaGraphManager
 
   private val methods = List(
-    Method(
+    CleanerMethod(
       "notMetaGraphContents",
       "Entities which do not exist in the meta-graph",
       "Truly orphan entities. Cached entities can get orphaned e.g. when the kite meta directory" +
         " is deleted or during a Kite version upgrade. Deleting these should not have any side" +
         " effects.",
       metaGraphContents),
-    Method(
+    CleanerMethod(
       "notReferredFromProjectTransitively",
       "Entities not associated with any project",
       "We consider an entity associated with a project if it's either directly referred to from" +
         " the project or it is used as an input to calculate another assoicated entity",
       transitivelyReferredFromProject),
-    Method(
+    CleanerMethod(
       "notReferredFromProject",
       "Entities not referenced by any project",
       "All entities except those that are vertex sets, edge bundles, attributes or scalars of" +
@@ -66,7 +67,7 @@ class CleanerController(environment: BigGraphEnvironment) {
   def getDataFilesStatus(user: serving.User, req: serving.Empty): DataFilesStatus = {
     assert(user.isAdmin, "Only administrator users can use the cleaner.")
     val files = getAllFiles()
-    val allFiles = files.entities ++ files.operations ++ files.scalars
+    val allFiles = files.partitioned ++ files.entities ++ files.operations ++ files.scalars
 
     DataFilesStatus(
       DataFilesStats(
@@ -79,6 +80,7 @@ class CleanerController(environment: BigGraphEnvironment) {
 
   private def getAllFiles(): AllFiles = {
     AllFiles(
+      getAllFilesInDir(io.PartitionedDir),
       getAllFilesInDir(io.EntitiesDir),
       getAllFilesInDir(io.OperationsDir),
       getAllFilesInDir(io.ScalarsDir))
@@ -87,13 +89,17 @@ class CleanerController(environment: BigGraphEnvironment) {
   // Return all files and dirs and their respective sizes in bytes in a
   // certain directory. Directories marked as deleted are not included.
   private def getAllFilesInDir(dir: String): Map[String, Long] = {
-    val hadoopFileDir = environment.dataManager.repositoryPath / dir
-    hadoopFileDir.listStatus.filterNot {
-      subDir => subDir.getPath().toString contains io.DeletedSfx
-    }.map { subDir =>
-      val baseName = subDir.getPath().getName()
-      baseName -> (hadoopFileDir / baseName).getContentSummary.getSpaceConsumed
-    }.toMap
+    val hadoopFileDir = environment.dataManager.writablePath / dir
+    if (!hadoopFileDir.exists) {
+      Map[String, Long]()
+    } else {
+      hadoopFileDir.listStatus.filterNot {
+        subDir => subDir.getPath().toString contains io.DeletedSfx
+      }.map { subDir =>
+        val baseName = subDir.getPath().getName()
+        baseName -> (hadoopFileDir / baseName).getContentSummary.getSpaceConsumed
+      }.toMap
+    }
   }
 
   private def metaGraphContents(): Set[String] = {
@@ -182,7 +188,9 @@ class CleanerController(environment: BigGraphEnvironment) {
     desc: String,
     filesToKeep: Set[String],
     allFiles: AllFiles): DataFilesStats = {
-    val filesToDelete = (allFiles.entities ++ allFiles.operations ++ allFiles.scalars) -- filesToKeep
+    val filesToDelete =
+      (allFiles.partitioned ++ allFiles.entities ++ allFiles.operations ++ allFiles.scalars) --
+        filesToKeep
     DataFilesStats(id, name, desc, filesToDelete.size, filesToDelete.map(_._2).sum)
   }
 
@@ -193,6 +201,7 @@ class CleanerController(environment: BigGraphEnvironment) {
     log.info(s"${user.email} attempting to mark orphan files deleted using '${req.method}'.")
     val files = getAllFiles()
     val filesToKeep = methods.find(m => m.id == req.method).get.filesToKeep()
+    markDeleted(io.PartitionedDir, files.partitioned.keys.toSet -- filesToKeep)
     markDeleted(io.EntitiesDir, files.entities.keys.toSet -- filesToKeep)
     markDeleted(io.OperationsDir, files.operations.keys.toSet -- filesToKeep)
     markDeleted(io.ScalarsDir, files.scalars.keys.toSet -- filesToKeep)
@@ -200,15 +209,18 @@ class CleanerController(environment: BigGraphEnvironment) {
 
   private def markDeleted(dir: String, files: Set[String]): Unit = {
     val hadoopFileDir = environment.dataManager.repositoryPath / dir
-    for (file <- files) {
-      (hadoopFileDir / file).renameTo(hadoopFileDir / (file + io.DeletedSfx))
+    if (hadoopFileDir.exists()) {
+      for (file <- files) {
+        (hadoopFileDir / file).renameTo(hadoopFileDir / (file + io.DeletedSfx))
+      }
+      log.info(s"${files.size} files marked deleted in ${hadoopFileDir.path}.")
     }
-    log.info(s"${files.size} files marked deleted in ${hadoopFileDir.path}.")
   }
 
   def deleteMarkedFiles(user: serving.User, req: serving.Empty): Unit = synchronized {
     assert(user.isAdmin, "Only administrators can delete marked files.")
     log.info(s"${user.email} attempting to delete marked files.")
+    deleteMarkedFilesInDir(io.PartitionedDir)
     deleteMarkedFilesInDir(io.EntitiesDir)
     deleteMarkedFilesInDir(io.OperationsDir)
     deleteMarkedFilesInDir(io.ScalarsDir)
@@ -216,11 +228,13 @@ class CleanerController(environment: BigGraphEnvironment) {
 
   private def deleteMarkedFilesInDir(dir: String): Unit = {
     val hadoopFileDir = environment.dataManager.repositoryPath / dir
-    hadoopFileDir.listStatus.filter {
-      subDir => subDir.getPath().toString contains io.DeletedSfx
-    }.map { subDir =>
-      (hadoopFileDir / subDir.getPath().getName()).delete()
+    if (hadoopFileDir.exists()) {
+      hadoopFileDir.listStatus.filter {
+        subDir => subDir.getPath().toString contains io.DeletedSfx
+      }.map { subDir =>
+        (hadoopFileDir / subDir.getPath().getName()).delete()
+      }
+      log.info(s"Deleted the marked files in ${hadoopFileDir.path}.")
     }
-    log.info(s"Deleted the marked files in ${hadoopFileDir.path}.")
   }
 }
